[FEAT] Move enum to enums folder
