+++ src/pages/UserProfile/Appointments/components/AppointmentDetail/AppointmentDetail.tsx
@@ -3,11 +3,10 @@ import { Link } from 'react-router-dom';
 import {
     AppointmentDetailProps,
     StatusConfig,
-    AppointmentStatus,
-    AppointmentType,
     getAppointmentTypeText as getTypeText,
     getAppointmentTypeIcon,
 } from '@/types/appointment.types';
+import { AppointmentStatus, AppointmentType } from '@/enums/appointment.enums';
 
 const AppointmentDetail: React.FC<AppointmentDetailProps> = ({
     appointment,
